# src/main/resources/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: JARAM Web API
  version: 0.1.0
  description: >
    JARAM(자람) 웹 백엔드 계약. FE 레포에서 파생된 단일 진실원이며 Spring Boot 백엔드가 구현한다.
    설계·usecase·결정 근거는 docs/superpowers/specs/2026-06-29-jaram-backend-design.md 참조.
    enum 와이어 값과 에러 코드는 FE 소비에 맞춰 고정되어 있다.
servers:
  - url: http://localhost:8080
    description: 로컬 개발 (Spring 기본 포트)
  # - url: https://api.jaram.example
  #   description: 운영 (배포 대상 확정 시 채움)
...
paths:
...
  /api/seminars:
    get:
      tags: [seminar]
      summary: 세미나 목록 조회 (UC-S1)
      description: status는 서버 파생. attendanceCode 미노출. public.
      responses:
        '200':
          description: 세미나 목록
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Seminar' }
        '5XX': { $ref: '#/components/responses/ServerError' }
    post:
      tags: [seminar]
      summary: 세미나 개설 (UC-S3)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SeminarCreateRequest' }
      responses:
        '201':
          description: 개설됨
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Seminar' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/Validation' }
        '5XX': { $ref: '#/components/responses/ServerError' }
  /api/seminars/{id}/attend:
    post:
      tags: [seminar]
      summary: 출석 체크 (UC-S2)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/SeminarId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AttendRequest' }
      responses:
        '200':
          description: 출석 완료 (중복 출석은 멱등)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttendResult' }
        '400':
          description: 코드 불일치 또는 출석 창 닫힘 (code=INVALID_CODE)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example: { code: INVALID_CODE, message: 출석 코드가 올바르지 않습니다. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '5XX': { $ref: '#/components/responses/ServerError' }
  /api/seminars/{id}/roster:
    get:
      tags: [seminar]
      summary: 출석 명단 조회 (UC-S4)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/SeminarId'
      responses:
        '200':
          description: 출석 명단
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RosterResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '5XX': { $ref: '#/components/responses/ServerError' }

  # ─────────────────────────── study ───────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Authorization: Bearer <accessToken>'
  parameters:
    SeminarId:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: 세미나 ID
...
  responses:
    Unauthorized:
      description: 인증 필요/토큰 무효 (FE 세션 clear)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: 권한 부족 (officer 전용)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: FORBIDDEN, message: 접근 권한이 없습니다. }
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Validation:
      description: 입력 검증 실패 (fieldErrors 포함)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            code: VALIDATION
            message: 입력값을 확인해 주세요.
            fieldErrors: { email: 한양대 이메일(@hanyang.ac.kr)만 사용할 수 있습니다. }
    ServerError:
      description: 서버 오류
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { code: SERVER, message: 일시적인 오류가 발생했습니다. }
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: 에러 코드 (FE가 분기에 사용)
        message: { type: string }
        fieldErrors:
          type: [object, 'null']
          additionalProperties: { type: string }
          description: 검증 실패 시 필드명 → 메시지
...
    SeminarStatus:
      type: string
      enum: [UPCOMING, ONGOING, ENDED]
      description: startsAt + 출석창으로 서버 파생
    TargetGrade:
      type: string
      enum: [NEWCOMER, ASSOCIATE, REGULAR, OB]
      description: 세미나 공개 대상 등급
...
    Seminar:
      type: object
      required: [id, title, startsAt, status, attendanceClosesAt]
      properties:
        id: { type: string }
        title: { type: string }
        speaker: { type: [string, 'null'] }
        topic: { type: [string, 'null'] }
        startsAt: { type: string, format: date-time, description: 정규 시작 시각 (ISO-8601) }
        day: { type: string, description: 파생 표시 (예 '27') }
        month: { type: string, description: 파생 표시 (예 '6월') }
        weekday: { type: string, description: 파생 표시 (예 '금') }
        time: { type: string, description: 파생 표시 (예 '19:00') }
        place: { type: [string, 'null'] }
        mode: { type: [string, 'null'] }
        status: { $ref: '#/components/schemas/SeminarStatus' }
        materialUrl: { type: [string, 'null'], format: uri }
        target:
          type: array
          items: { $ref: '#/components/schemas/TargetGrade' }
          description: 공개 대상 등급. 빈 배열은 전체 공개.
        description: { type: [string, 'null'], description: 세미나 상세 설명 }
        attendanceClosesAt: { type: string, format: date-time, description: '출석 인정 마감 시각 (startsAt + 출석창, 서버 파생)' }
        attendedAt: { type: [string, 'null'], description: '내 출석 시각 표시(예 19:02). 미출석/비로그인이면 null' }
    SeminarCreateRequest:
      type: object
      required: [title, startsAt]
      properties:
        title: { type: string }
        speaker: { type: [string, 'null'] }
        topic: { type: [string, 'null'] }
        startsAt: { type: string, format: date-time }
        place: { type: [string, 'null'] }
        mode: { type: [string, 'null'] }
        attendanceCode: { type: [string, 'null'], description: 출석 코드 (응답엔 미노출) }
        materialUrl: { type: [string, 'null'], format: uri }
        target:
          type: array
          items: { $ref: '#/components/schemas/TargetGrade' }
          description: 공개 대상 등급. 빈 배열은 전체 공개.
        description: { type: [string, 'null'] }
    AttendRequest:
      type: object
      required: [code]
      properties:
        code: { type: string }
    AttendResult:
      type: object
      required: [seminarId, at]
      properties:
        seminarId: { type: string }
        at: { type: string, description: 출석 시각 표시 (예 '19:02') }
    RosterEntry:
      type: object
      required: [name, sid, at]
      properties:
        name: { type: string }
        sid: { type: string, description: 학번 }
        at: { type: string, description: 출석 시각 표시 (예 '19:02') }
    RosterResponse:
      type: object
      required: [title, cap, list]
      properties:
        title: { type: string }
        cap: { type: integer, description: 정원 }
        list:
          type: array
          items: { $ref: '#/components/schemas/RosterEntry' }

    # ── study ──
...